feat: extract shared foundry ExpandEnv into core pkg/foundry
