+++ cli/azd/pkg/foundry/templating_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package project
+package foundry
 
 import (
 	"testing"
